Refactor: chattingForm 분리 (#62)
